chore(nest-cli.json): add nest-cli.json configuration file to the project
The nest-cli.json file is added to the project. This file contains the configuration for the NestJS CLI. It specifies the schema, collection, source root, entry file, and compiler options for the project. The "deleteOutDir" option is set to true, which means that the output directory will be deleted before each build. This configuration file is necessary for the proper functioning of the NestJS CLI.