feat(optional.ts): add `Optional` type to make specified properties optional on a given type
The `Optional` type allows making specific properties optional on a given type. This can be useful when you want to create a new type with some properties being optional while keeping the rest of the properties intact.

Example usage:
```typescript
type Post {
  id: string;
  name: string;
  email: string;
}

Optional<Post, 'id' | 'email'>
```

In the above example, the `Optional` type is used to make the `id` and `email` properties optional on the `Post` type.