feat(entity.ts): add Entity base class to provide common functionality for all entities in the application
The Entity base class is added to provide common functionality for all entities in the application. It includes a private `_id` property of type `UniqueEntityID` to uniquely identify each entity, and a `props` property of type `Props` to store the entity's properties.

The `id` getter is implemented to retrieve the `_id` property.

The constructor of the Entity class takes in `props` of type `Props` and an optional `id` of type `UniqueEntityID`. If `id` is not provided, a new `UniqueEntityID` is generated and assigned to `_id`.