feat(errors): add IUseCaseError interface to define common structure for use case errors
The commit adds a new file `use-case.error.ts` which defines an interface `IUseCaseError`. This interface has a single property `message` of type string. This interface will be used to define the structure of errors that can occur in use cases. By having a common structure for use case errors, it will be easier to handle and manage errors consistently across the application.