feat(audio): add InvalidAudioTypeError class to handle errors related to invalid audio file types
The InvalidAudioTypeError class is added to the audio domain's use-cases/errors directory. This class extends the Error class and implements the IUseCaseError interface. It takes a type parameter in its constructor and sets the error message to indicate that the provided file type is not valid. This class will be used to handle errors related to invalid audio file types in the application.