feat(presenters): add PromptPresenter class to convert PromptEntity to HTTP response format
The PromptPresenter class is added to the presenters directory in the infra/http module. This class provides a static method `toHTTP` that takes a PromptEntity object as input and returns an object in the format required for the HTTP response. The properties of the PromptEntity object are mapped to the corresponding properties in the response object.